Fixed pointer conversion warnings.
git-svn-id: https://cgdb.svn.sourceforge.net/svnroot/cgdb/cgdb/trunk@884 61d65b3c-fd13-0410-b770-9cb5f9339520

+++ various/adt/src/std_hash_driver.c
@@ -292,7 +292,7 @@ static int find_first     (void* key,
 
 static void direct_hash_test (void)
 {
-     int       i, rc;
+     long       i, rc;
      struct std_hashtable     *h;
 
      h = std_hash_table_new (NULL, NULL);
@@ -307,7 +307,7 @@ static void direct_hash_test (void)
 
      for (i=1; i<=20; i++)
           {
-          rc = ((int) (
+          rc = ((long) (
 	  	std_hash_table_lookup (h, ((void*) (i)))));
 
 	  assert (rc != 0);
